WIP: array operators and CUDAArray class
