+++ pyclesperanto_prototype/_tier0/_backends.py
@@ -1,2 +1,5 @@
 from ._opencl_backend import opencl_backend
 _current_backend = opencl_backend()
+
+#from ._cuda_backend import cuda_backend
+#_current_backend = cuda_backend()
